Restructure mtdev api
Split the api into plumbing and porcelain layers and move the
plumbing part to its own optional header file.

The main usecase is to fetch events from the device, route them
through the converter and extract the processed events. To simplify
the API, replace the intermediate mtdev_pull() function by the
higher-level mtdev_get(). This function does all the required steps,
and has the same semantics as read().

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>